Mouse events implemented
Currently program displays all the cards in rows and when you click on a card, it switches between the card being dark and light.
Next, I will have to start programming the logic behind solitaire itself.
In order to get the cards overlapping successfully, I will have to make a method for changing the order of the cards arraylist in CardDisplay based on where the cards are. I may also make a private class called column for each column in the game.